Russian translation
Note that the config reload button text (`text.owo.config.button.reload`) does not fit the reload button, so I had to contract it. You might want to fix that in the mod so that the button gets its size from its content or something like that, I'm not that experienced in the UI part of the mod yet.